Add create estimate from existing estimate
There is an end point to create an estimate
from an existing estimate. I needed to add this
end point to the API class so that I could
have this functionality in my project.

The need arose from a requirement to quickly
display an estimated cost for a product and some
products only have the option to use an
existing estimate as a template.