Get Product By ID
There is a need to be able to use the products
endpoint and use a product ID to get a specific
product.

I added a new function to the api class that
allows for this functionality. The reason I did
not remake the existing product function is
because it would be a breaking change that
would affect other repositories that use
this library.